break out AtmoUserSerializer
