+++ api/serializers/__init__.py
@@ -26,3 +26,4 @@ from .post_license_serializer import POST_LicenseSerializer
 from .machine_request_serializer import MachineRequestSerializer
 from .maintenance_record_serializer import MaintenanceRecordSerializer
 from .identity_detail_serializer import IdentityDetailSerializer
+from .atmo_user_serializer import AtmoUserSerializer
